Fix package name and move file
